smart-ocr-saas/backend/requirements.txt
# FastAPI & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
motor>=3.3.0
pymongo>=4.6.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
ldap3>=2.9.0

# OpenAI
openai>=1.3.0

# MinIO
minio>=7.2.0

# PDF Processing
PyMuPDF>=1.23.0

# Excel Export
openpyxl>=3.1.0

# Redis
redis>=5.0.0

# Settings
pydantic-settings>=2.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
black>=23.11.0
isort>=5.12.0
